requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-crispy-forms
crispy-bootstrap5